+++ MyCourse/Models/Exceptions/CourseNotFoundException.cs
@@ -4,7 +4,8 @@ namespace MyCourse.Models.Exceptions
 {
     public class CourseNotFoundException : Exception
     {
-        public CourseNotFoundException(int courseId) : base($"Course {courseId} not found!" )
+        //public CourseNotFoundException(long courseId) : base($"Course {courseId} not found!")
+        public CourseNotFoundException(long courseId) : base($"Course {courseId} not found!")
         {
         }
     }
